Move filter related tests to filtering
